// src/core/plugins/core-factions/shared/interfaces.ts
type _id = string;

export const RankPermissionNames = {
    bankAdd: 'bankAdd',
    bankRemove: 'bankRemove',
    manageRanks: 'manageRanks',
    manageRankPermissions: 'manageRankPermissions',
    manageMembers: 'manageMembers',
    kickMembers: 'kickMembers',
    addMembers: 'addMembers',
};

/**
 * Permissions that belong to an individual rank.
 * @export
 * @interface RankPermissions
 */
export interface RankPermissions {
    /**
     * Can this rank add to the bank?
     *
     * @type {boolean}
     * @memberof FactionRank
     */
    bankAdd: boolean;

    /**
     * Can this rank remove from the bank,
     * purchase vehicles,
     * and do general faction bank purchases.
     *
     * @type {boolean}
     * @memberof FactionRank
     */
    bankRemove: boolean;

    /**
     * Can this rank manage ranks lower than itself?
     *
     * @type {boolean}
     * @memberof FactionRank
     */
    manageRanks: boolean;

    /**
     * Can this rank manage rank permissions lower than itself?
     *
     * @type {boolean}
     * @memberof RankPermissions
     */
    manageRankPermissions: boolean;

    /**
     * Can this rank manage members by promoting and demoting lower than itself?
     *
     * @type {boolean}
     * @memberof FactionRank
     */
    manageMembers: boolean;

    /**
     * Can this rank kick members lower than itself?
     *
     * @type {boolean}
     * @memberof FactionRank
     */
    kickMembers: boolean;

    /**
     * Can this rank add members lower than itself?
     *
     * @type {boolean}
     * @memberof RankPermissions
     */
    addMembers: boolean;

    /**
     * Can manage vehicles by purchasing vehicles available.
     *
     * @type {boolean}
     * @memberof RankPermissions
     */
    manageVehicles?: boolean;
}
// ...
export interface FactionRank {
    [key: string]: any;

    /**
     * A unique identifier for the Faction Rank
     * Should be auto-generated.
     *
     * @type {string}
     * @memberof FactionRank
     */
    uid?: string;

    /**
     * The name of this Rank
     *
     * @type {string}
     * @memberof FactonRank
     */
    name: string;

    /**
     * Custom action permissions this rank has access to.
     *
     * @type {Array<string>}
     * @memberof FactionRank
     */
    actionPermissions: Array<string>;

    /**
     * Permissions for the rank.
     *
     * @type {RankPermissions}
     * @memberof FactionRank
     */
    rankPermissions: RankPermissions;

    /**
     * Vehicles this rank has keys to by default.
     *
     * @type {Array<_id>}
     * @memberof FactionRank
     */
    vehicles: Array<_id>;

    /**
     * The higher the weight the more important they are.
     * Owner rank is always 99.
     * Nothing can exceed 99.
     *
     * @type {number}
     * @memberof FactionRank
     */
    weight: number;
}
// ...
export interface FactionSettings {
    [key: string]: any;

    /**
     * The current positions of faction headquarters.
     *
     * @type {Vector3}
     * @memberof FactionSettings
     */
    position?: Vector3;

    /**
     * A list of vehicle parking spots to spawn vehicles from.
     *
     * @type {Array<Vector3>}
     * @memberof Faction
     */
    parkingSpots?: Array<{ pos: Vector3; rot: Vector3 }>;

    /**
     * A list of vehicles available for purchase for this faction.
     *
     * @type {Array<{ model: string, price: number }>}
     * @memberof FactionSettings
     */
    vehicles?: Array<{ model: string; price: number }>;

    /**
     * Maximum number of vehicles purchaseable by the Faction
     *
     * @type {number}
     * @memberof FactionSettings
     */
    maxVehicles?: number;

    /**
     * A blip sprite index.
     *
     * @type {number}
     * @memberof FactionSettings
     */
    blip?: number;

    /**
     * A blip sprite colour.
     *
     * @type {number}
     * @memberof FactionSettings
     */
    blipColor?: number;

    /**
     * Message of the day for factions.
     *
     * @type {string}
     * @memberof FactionSettings
     */
    motd?: string;
}